OpenAI Agents SDK/basics_agents/myapp/requirements.txt
openai-agents>=0.0.16
python-dotenv>=0.9.9 